[test] Fix WeakGlobalHandle test with --stress-concurrent-allocation
Use Global instead of Persistent such that GlobalHandle is reset at the
end of the function. Persistent doesn't reset in the destructor,
which means that the GC resets the GlobalHandle. With
--stress-concurrent-allocation this might not happen in the test
function itself but when the cctest framework itself works through
the event queue. At that point the Persistent isn't live anymore.

Bug: v8:10315
Change-Id: If77388ad5acb80538852beca0ab22a4ebaf0b5c1