[nci] Prepare JSForInPrepare and JSForInNext for feedback input
These two operators are still missing feedback collection in generic
lowering (reminder: all operations that collect FB in the interpreter
must also collect FB in generic lowering).

This CL prepares for that by adding the feedback vector as an input,
and additionally adds node wrappers to improve useability.

The actual collection logic will be added in a following CL.

Bug: v8:8888
Change-Id: I04627eedb2dc237dc4e417091c44d2a95bd98f5f