[regexp] Increase number of backtracks before fallback
This commit increases the number of backtracks before we fall back to
the experimental engine.  The previous limit was reached just before
irregexp finished in sunspider/string-tagcloud and
web-tooling-benchmark/coffeescript.  This meant that we aborted
execution and started over with a slower (but asymptotically faster)
engine just before we would've finished, causing a performance
regression.

Bug: chromium:1138975
Bug: v8:10765
Change-Id: I10a4b149300f88680084604310730ee91758bfa0