[x64] Introduce helpers to print xmm operands
XMM registers/operands can be mixed with normal operands. We introduce 4
new enums to handle this case:

- xmmreg, oper
- oper, xmmreg
- xmmreg, xmmoper
- xmmoper, xmmreg

The different between op and xmmop is what type of register to print if
the modrm byte ends up decoding into a register.

The only difference between the output of the DisasmX64 test is pinsrd,
pinsrq, and pshuflw which had extras spaces. This is fixed with this
patch.

Test: cctest test-disasm-x64/DisasmX64 --random-seed=-1547905187

Bug: v8:10933
Change-Id: I2904c0fe83579e8df83ff2f304409278f1874c2d