Avoid pushing/popping lr in bytecode handlers
This patch changes the way bytecode handler frames are constructed,
instead of pushing the lr at the beginning of each bytecode handler
and popping it at the end, the lr is pushed right after
InterpreterEntryTrampoline and just before entering the first handler.
This reduces the number of times the lr is signed and authenticated in
cfi builds, and reduces as well the size of the builtins by 8256 bytes
in cfi builds at the cost of 224 extra bytes in non-cfi builds.
There is no change regarding the fp.

What happens is that before calling the first bytecode handler, the
InterpreterEntryTrampoline calls InterpreterBytecodeHandlerTrampoline
that will sign and store the LR along with some padding on the stack,
and then call the first bytecode. Each bytecode then normally stores
and restores the fp in the remaining padding slot if needed. This
implies that even when a bytecode does not have a frame, there will be
two extra slots containing the LR (and some padding/old fp). This
requires a few changes to the way the sp is adjusted before a tail
call, to prevent the LR from being dropped.
The LR is restored only when returning to the interpreter, through the
return bytecode or a tail call to a builtin and a dispatch.

Change-Id: Id34d87f67cbc607a4313478de360450c8c0cd3c2