[class] Load the super constructor after evaluating arguments
Even though the specification calls GetSuperConstructor before
evaluating arguments, the only observable thing is the timing of the
error if that constructor is not actually a constructor. Namely, that
the error must be thrown after a super call's arguments are evaluated.
This CL moves the GetSuperConstructor bytecode to be after argument
evaluation.

https://tc39.es/ecma262/#sec-super-keyword

Bug: v8:10111
Change-Id: I3af99e32a34d99493806bb01b547d6f671cdc9de