[wasm] Bump maximum code space to 4GB on 64-bit systems
Big partners are still hitting the code space limit, especially when
debugging larger wasm apps. The limits was bumped to 2GB in August, but
we already have security clearance to increase it to 4GB.
On 32-bit systems, a limit of 4GB is problematic, and in fact we would
run into integer overflows if we try to set such a high limit. It also
does not make much sense, since in a 4GB address space it's highly
unlikely that we are able to allocate more than 2GB for wasm code
anyway.
Thus, this CL only bumps the limit for 64-bit systems, and keeps the
limit at 2GB for 32-bit systems.

R=ecmziegler@chromium.org

Bug: chromium:1143118
Change-Id: I94e40cbf2da5c24f2ed874f95f208c22500b5b10