[csa][cleanup] TNodify StringAdd_CheckNone buitin
StringAdd_CheckNone is called from Turbofan with an empty context. This
builtin needs context when calling the StringAdd runtime function which
could potentially throw. Turbofan does bounds check before calling this
builtin so it is safe to pass an empty context. To enable TNodification
of this builtin this cl adds a new type that either accepts a context
or an empty context (Smi::Zero) and updates the builtin to use this new
type.

Bug: v8:6949, v8:11074
Change-Id: Iff12b391ff95109649f2c81fe081e277850f60d6