[dict-proto] Preparing ordered hash tables for usage as property dicts
This contains more changes to ordered hash tables towards using them as
property dictionaries.
Most notably, this CL makes the type of the used isolates a template
parameter for certain operations. This is already the case for
unordered hash tables, and necessary in follow-up CLs where ordered
name dictionaries are used with LocalIsolate as the isolate type.

Bug: v8:7569
Change-Id: I5c938425a2c196ccd0866b66318a350ebeac8be2